pyproject.toml
[tool.poetry]
name = "bobai"
version = "0.1.0"
description = ""
authors = ["sharp119 <cs.aditya58@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-generativeai = "^0.7.1"
langchain-google-genai = "^1.0.7"
langchain = "^0.2.6"
pymysql = "^1.1.1"
langchain-experimental = "^0.0.62"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
